[assets] fix infinite loop upon etcd compaction.
Assets monitors etcd changes statefully by storing the last watched
revision in the filesystem.  When etcd compacted the database while
sabakan is stopping, sabakan would not be able to resume the stateful
watch from the saved revision.

In this case, sabakan resets the revision stored in the filesystem
to zero and re-synchronize all assets with the current database.

Unfortunately, sabakan could generate corrupted revision file
because it did not truncate the revision file.  This commit
truncates the revision file properly, and adds unit tests for them.