Move gateway offset from dhcp.json to ipam.json
"gateway-offset" field in dhcp.json was used to calculate
the default gateway address in a subnet.

Because the subnet is also calculated from ipam.json,
placing "gateway-offset" in DHCP configurations is considered
a design flaw.

This PR adds "node-gateway-offset" field to ipam.json and
uses it to calculate the default gateway addresses.